Define test/install dependencies.
We use py.test for testing. Tell what packages we need in setup.py, so
that we can run `python setup.py test` to install and run tests.

We deploy the manual py.test resgistration in setup.py, because the
alternative "integrated" approach requires to set the setuptools
`setup_requires` var, which consider to be bad. Instead we try to avoid
using `setup_requires`.