Move HTML theme into source tree.
Locations inside some package are the only ones that we can ship easily
with the Python code and that are more or less guaranteed to be
reachable after install process.

Unfortunately, data files stored elsewhere (outside of source trees),
are not neccessarily stored in well known locations by different
packagemanagers (`distutils`, `setuptools`, etc.).